Refactor csdl2openapi with modularized functions and improved code structure
