+++ package.json
@@ -38,6 +38,7 @@
     "@sap/cds": ">=7.6"
   },
   "devDependencies": {
+    "@sap/cds": "^9.2.1",
     "@types/jest": "^30.0.0",
     "@types/node": "^24.3.0",
     "@types/pluralize": "^0.0.33",
